requirements.txt
# ---------------------------------------------------------------------------
# Environment: Python 3.10.x + NVIDIA GPU (CUDA). Install a matching PyTorch
# build *before* this file — see https://pytorch.org/get-started/locally/
#
# CUDA extensions (from repository root, after PyTorch is available):
#   pip install ./submodules/simple-knn
#   pip install ./submodules/diff-surfel-rasterization
#   pip install ./utils/pointnet_lib
#
# SAM ViT-H weights (required for utils/get_mask.py and stage 1):
#   wget -O submodules/sam_vit_h_4b8939.pth \
#     https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth
#
# PyTorch3D: wheels often depend on exact torch/CUDA; if pip fails, follow:
#   https://github.com/facebookresearch/pytorch3d/blob/main/INSTALL.md
# tiny-cuda-nn: requires nvcc (CUDA toolkit); see package docs if build fails.
# ---------------------------------------------------------------------------

git+https://github.com/facebookresearch/segment-anything.git

joblib==1.4.2
matplotlib==3.10.8
mediapy==1.2.6
numpy==2.4.4
open3d==0.19.0
opencv-python==4.10.0.84
pandas==3.0.2
Pillow==12.2.0
piq==0.8.0
plyfile==1.1.3
pytorch3d==0.7.8
pytorch_lightning==2.4.0
scikit-image==0.24.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==75.8.0
tinycudann==1.7
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
trimesh==4.5.1
